Keep us in the present
Fixes forum sunrise bug, where we would be living in the future but
posting in the past between 12am - 8am UTC every day. This is because I
was calculating the sunrise for _today_, so when UTC's notion of today
changed (eg after midnight), it caused a problem with our notion of
present.